feat: complete S3 integration and cleanup unused dependencies
This commit completes the S3 bucket integration for storing static and media files, as outlined in the README.md. It also removes unnecessary dependencies such as django-vite, django_components, loguru, PyYaml, Brotli, colorama and whitenoise, resulting in a more streamlined and efficient project.

The settings.py file was modified to only apply the s3 configuration when DEBUG=False.